src/requirements.txt
pycrypto
requests
urllib3
colorlog
PyYAML
bitmath
pyOpenSSL
dnspython
pythonwhois==2.2.2
shodan
scandir
pyravendb
progressbar
retry